perf(app-router): defer cache-only runtimes
